Implement profile management, admin dashboard, and fake backend

+++ frontend/src/app/admin/accounts/list.component.ts
@@ -1,12 +1,12 @@
 import { Component, OnInit } from '@angular/core';
 import { first } from 'rxjs/operators';
 
-import { AccountService } from '@app/_services';
-import { Account } from '@app/_models';
+import { AccountService } from '../../_services/account.service';
+import { Account } from '../../_models/account';
 
 @Component({ templateUrl: 'list.component.html' })
 export class ListComponent implements OnInit {
-    accounts: any[];
+    accounts: Account[] = [];
 
     constructor(private accountService: AccountService) {}
 
@@ -18,11 +18,12 @@ export class ListComponent implements OnInit {
 
     deleteAccount(id: string) {
         const account = this.accounts.find(x => x.id === id);
+        if (!account) return;
         account.isDeleting = true;
         this.accountService.delete(id)
             .pipe(first())
             .subscribe(() => {
-                this.accounts = this.accounts.filter(x => x.id !== id);
+                this.accounts = this.accounts.filter(x => x.id !== id)
             });
     }
 }
